+++ tests/test_skool2ctl.py
@@ -1,10 +1,14 @@
-import textwrap
+from textwrap import dedent
 from unittest.mock import patch
 
 from skoolkittest import SkoolKitTestCase
 from skoolkit import skool2ctl, VERSION
+from skoolkit.config import COMMANDS
 
 ELEMENTS = 'abtdrmscn'
+
+def mock_config(name):
+    return {k: v[0] for k, v in COMMANDS[name].items()}
 
 class MockCtlWriter:
     def __init__(self, skoolfile, elements, write_hex, preserve_base, min_address, max_address, keep_lines):
@@ -23,12 +27,12 @@ class MockCtlWriter:
         self.write_called = True
 
 class Skool2CtlTest(SkoolKitTestCase):
-    def _check_ctl_writer(self, skoolfile, elements=ELEMENTS, write_hex=0, preserve_base=False,
+    def _check_ctl_writer(self, skoolfile, elements=ELEMENTS, write_hex=0, preserve_base=0,
                           min_address=0, max_address=65536, keep_lines=0):
         self.assertEqual(mock_ctl_writer.skoolfile, skoolfile)
         self.assertEqual(mock_ctl_writer.elements, elements)
         self.assertEqual(mock_ctl_writer.write_hex, write_hex)
-        self.assertIs(mock_ctl_writer.preserve_base, preserve_base)
+        self.assertEqual(mock_ctl_writer.preserve_base, preserve_base)
         self.assertEqual(mock_ctl_writer.min_address, min_address)
         self.assertEqual(mock_ctl_writer.max_address, max_address)
         self.assertEqual(mock_ctl_writer.keep_lines, keep_lines)
@@ -46,7 +50,32 @@ class Skool2CtlTest(SkoolKitTestCase):
             self.assertTrue(error.startswith('usage: skool2ctl.py'))
 
     @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
+    @patch.object(skool2ctl, 'get_config', mock_config)
     def test_default_option_values(self):
+        skoolfile = 'test.skool'
+        skool2ctl.main((skoolfile,))
+        self._check_ctl_writer(skoolfile)
+
+    @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
+    def test_config_read_from_file(self):
+        ini = """
+            [skool2ctl]
+            Hex=1
+            KeepLines=1
+        """
+        self.write_text_file(dedent(ini).strip(), 'skoolkit.ini')
+        skoolfile = 'test.skool'
+        skool2ctl.main((skoolfile,))
+        self._check_ctl_writer(skoolfile, write_hex=1, keep_lines=1)
+
+    @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
+    def test_invalid_option_values_read_from_file(self):
+        ini = """
+            [skool2ctl]
+            Hex=?
+            KeepLines=x
+        """
+        self.write_text_file(dedent(ini).strip(), 'skoolkit.ini')
         skoolfile = 'test.skool'
         skool2ctl.main((skoolfile,))
         self._check_ctl_writer(skoolfile)
@@ -76,21 +105,21 @@ class Skool2CtlTest(SkoolKitTestCase):
         skoolfile = 'test.skool'
         for option in ('-h', '--hex'):
             skool2ctl.main((option, skoolfile))
-            self._check_ctl_writer(skoolfile, write_hex=1)
+            self._check_ctl_writer(skoolfile, write_hex=2)
 
     @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
     def test_option_l(self):
         skoolfile = 'test.skool'
         for option in ('-l', '--hex-lower'):
             skool2ctl.main((option, skoolfile))
-            self._check_ctl_writer(skoolfile, write_hex=-1)
+            self._check_ctl_writer(skoolfile, write_hex=1)
 
     @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
     def test_option_b(self):
         skoolfile = 'test.skool'
         for option in ('-b', '--preserve-base'):
             skool2ctl.main((option, skoolfile))
-            self._check_ctl_writer(skoolfile, preserve_base=True)
+            self._check_ctl_writer(skoolfile, preserve_base=1)
 
     @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
     def test_option_S(self):
@@ -128,12 +157,77 @@ class Skool2CtlTest(SkoolKitTestCase):
         skool2ctl.main(('-S', str(start), '-E', str(end), skoolfile))
         self._check_ctl_writer(skoolfile, min_address=start, max_address=end)
 
+    @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
+    @patch.object(skool2ctl, 'get_config', mock_config)
+    def test_option_I(self):
+        self.run_skool2ctl('-I Hex=1 test.skool')
+        self._check_ctl_writer('test.skool', write_hex=1)
+
+    @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
+    @patch.object(skool2ctl, 'get_config', mock_config)
+    def test_option_I_overrides_other_options(self):
+        self.run_skool2ctl('-h -I Hex=1 test.skool')
+        self._check_ctl_writer('test.skool', write_hex=1)
+
+    @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
+    def test_option_I_overrides_config_read_from_file(self):
+        ini = """
+            [skool2ctl]
+            Hex=2
+            KeepLines=1
+        """
+        self.write_text_file(dedent(ini).strip(), 'skoolkit.ini')
+        self.run_skool2ctl('--ini Hex=1 -I KeepLines=0 test.skool')
+        self._check_ctl_writer('test.skool', write_hex=1, keep_lines=0)
+
+    @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
+    @patch.object(skool2ctl, 'get_config', mock_config)
+    def test_option_I_multiple(self):
+        self.run_skool2ctl('-I Hex=1 --ini PreserveBase=1 test.skool')
+        self._check_ctl_writer('test.skool', write_hex=1, preserve_base=1)
+
+    @patch.object(skool2ctl, 'CtlWriter', MockCtlWriter)
+    @patch.object(skool2ctl, 'get_config', mock_config)
+    def test_option_I_invalid_value(self):
+        self.run_skool2ctl('-I Hex=x test.skool')
+        self._check_ctl_writer('test.skool')
+
+    @patch.object(skool2ctl, 'get_config', mock_config)
+    def test_option_show_config(self):
+        output, error = self.run_skool2ctl('--show-config', catch_exit=0)
+        self.assertEqual(error, '')
+        exp_output = """
+            [skool2ctl]
+            Hex=0
+            KeepLines=0
+            PreserveBase=0
+        """
+        self.assertEqual(dedent(exp_output).strip(), output.rstrip())
+
+    def test_option_show_config_read_from_file(self):
+        ini = """
+            [skool2ctl]
+            Hex=1
+            KeepLines=1
+            PreserveBase=1
+        """
+        self.write_text_file(dedent(ini).strip(), 'skoolkit.ini')
+        output, error = self.run_skool2ctl('--show-config', catch_exit=0)
+        self.assertEqual(error, '')
+        exp_output = """
+            [skool2ctl]
+            Hex=1
+            KeepLines=1
+            PreserveBase=1
+        """
+        self.assertEqual(dedent(exp_output).strip(), output.rstrip())
+
     def test_run(self):
         skool = """
             ; Test skool file for skool2ctl testing
             c65535 RET
         """
-        skoolfile = self.write_text_file(textwrap.dedent(skool).strip(), suffix='.skool')
+        skoolfile = self.write_text_file(dedent(skool).strip(), suffix='.skool')
         output, error = self.run_skool2ctl(skoolfile)
         self.assertEqual(error, '')
         self.assertEqual(output, 'c 65535 Test skool file for skool2ctl testing\n')